Namespace for agent features tests
