+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/AutostartDisabled.cs
@@ -7,7 +7,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class AutoStartDisabled : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
