+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/CallStackFallbackMvc.cs
@@ -9,7 +9,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class CallStackFallbackMvc : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
