+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/ConnectResponseHandlingTests.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class ConnectResponseHandlingTests : IClassFixture<MvcWithCollectorFixture>
