+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/DataTransmissionPutGzip.cs
@@ -11,7 +11,7 @@ using System.Linq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class DataTransmissionPutGzip : IClassFixture<MvcWithCollectorFixture>
