+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/GetBrowserTimingHeader.cs
@@ -2,24 +2,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
 
-using System;
 using System.Collections.Generic;
 using NewRelic.Agent.IntegrationTestHelpers;
 using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
-    public class GetBrowserTimingHeaderAutoOn : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
+    public class GetBrowserTimingHeader : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
     {
+
         private readonly RemoteServiceFixtures.BasicMvcApplicationTestFixture _fixture;
 
         private string _browserTimingHeader;
-        private string _htmlContentAfterCallToGetBrowserTiming;
 
-        public GetBrowserTimingHeaderAutoOn(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper output)
+        public GetBrowserTimingHeader(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -30,14 +29,12 @@ namespace NewRelic.Agent.IntegrationTests
                     var configPath = fixture.DestinationNewRelicConfigFilePath;
 
                     var configModifier = new NewRelicConfigModifier(configPath);
-                    configModifier.AutoInstrumentBrowserMonitoring(true);
+                    configModifier.AutoInstrumentBrowserMonitoring(false);
                     configModifier.BrowserMonitoringEnableAttributes(true);
-
                 },
                 exerciseApplication: () =>
                 {
                     _browserTimingHeader = _fixture.GetBrowserTimingHeader();
-                    _htmlContentAfterCallToGetBrowserTiming = _fixture.GetHtmlWithCallToGetBrowserTimingHeader();
                 }
             );
             _fixture.Initialize();
@@ -46,10 +43,7 @@ namespace NewRelic.Agent.IntegrationTests
         [Fact]
         public void Test()
         {
-            NrAssert.Multiple(
-                () => Assert.Contains("NREUM", _browserTimingHeader),
-                ShouldNotAutoInstrumentAfterCallToGetBrowserTimingHeader
-            );
+            Assert.Contains("NREUM", _browserTimingHeader);
 
             var browserMonitoringConfig = JavaScriptAgent.GetJavaScriptAgentConfigFromSource(_browserTimingHeader);
 
@@ -64,7 +58,6 @@ namespace NewRelic.Agent.IntegrationTests
                 () => Assert.Contains("agent", browserMonitoringConfig.Keys),
                 () => Assert.Contains("atts", browserMonitoringConfig.Keys)
             );
-
             var attrsDict = HeaderEncoder.DecodeAndDeserialize<Dictionary<string, IDictionary<string, object>>>(browserMonitoringConfig["atts"], _fixture.TestConfiguration.LicenseKey, 13);
             Assert.Contains("a", attrsDict.Keys);
             IDictionary<string, object> agentAttrsDict = attrsDict["a"];
@@ -80,11 +73,6 @@ namespace NewRelic.Agent.IntegrationTests
                 () => Assert.NotNull(browserMonitoringConfig["applicationTime"]),
                 () => Assert.NotNull(browserMonitoringConfig["agent"])
             );
-        }
-
-        private void ShouldNotAutoInstrumentAfterCallToGetBrowserTimingHeader()
-        {
-            Assert.DoesNotContain("NREUM", _htmlContentAfterCallToGetBrowserTiming);
         }
     }
 }
