+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/MemoryMvc.cs
@@ -10,18 +10,17 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
-    // This test verifies that all supportability metrics are generated that are required by APM.
     [NetFrameworkTest]
-    public class RequiredSupportabilityMetrics : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
+    public class MemoryMvc : IClassFixture<RemoteServiceFixtures.BasicMvcApplicationTestFixture>
     {
         private readonly RemoteServiceFixtures.BasicMvcApplicationTestFixture _fixture;
 
-        public RequiredSupportabilityMetrics(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper output)
+        public MemoryMvc(RemoteServiceFixtures.BasicMvcApplicationTestFixture fixture, ITestOutputHelper testLogger)
         {
             _fixture = fixture;
-            _fixture.TestLogger = output;
+            _fixture.TestLogger = testLogger;
             _fixture.Actions
             (
                 setupConfiguration: () =>
@@ -31,14 +30,11 @@ namespace NewRelic.Agent.IntegrationTests
                 exerciseApplication: () =>
                 {
                     _fixture.Get();
-                    _fixture.Get();
-                    _fixture.Get();
-                    _fixture.Get();
-
-                    // The Supportability/AnalyticsEvents/TotalEventsSent metric won't be seen until a second harvest occurs, so we must wait up to 60 seconds for it
                     var startTime = DateTime.Now;
-                    while (DateTime.Now <= startTime.AddSeconds(60) && !_fixture.AgentLog.GetMetrics().Any(metric => metric.MetricSpec.Name == "Supportability/AnalyticsEvents/TotalEventsSent"))
+                    while (DateTime.Now <= startTime.AddSeconds(60))
                     {
+                        if (_fixture.AgentLog.GetMetrics().Any(metric => metric.MetricSpec.Name == "Memory/Physical"))
+                            break;
                         Thread.Sleep(TimeSpan.FromSeconds(5));
                     }
                 }
@@ -51,9 +47,7 @@ namespace NewRelic.Agent.IntegrationTests
         {
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
-                new Assertions.ExpectedMetric { metricName = @"Supportability/MetricHarvest/transmit", callCount = 1 },
-                new Assertions.ExpectedMetric { metricName = @"Supportability/AnalyticsEvents/TotalEventsSeen", callCount = 4 },
-                new Assertions.ExpectedMetric { metricName = @"Supportability/AnalyticsEvents/TotalEventsSent", callCount = 4 },
+                new Assertions.ExpectedMetric {metricName = @"Memory/Physical"}
             };
 
             var metrics = _fixture.AgentLog.GetMetrics().ToList();
