+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/RemotingSerialization.cs
@@ -6,7 +6,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class RemotingSerialization : IClassFixture<RemoteServiceFixtures.OwinRemotingFixture>
