+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/ThreadProfileNetCoreTests.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetCoreTest]
     public class ThreadProfileNetCoreTests : IClassFixture<AspNetCoreMvcWithCollectorFixture>
