+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/ThreadProfileStressTests.cs
@@ -10,7 +10,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class ThreadProfileStressTests : IClassFixture<ThreadProfileStressTestWithCollectorFixture>
