+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/ThreadProfileTests.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.AgentFeatures
 {
     [NetFrameworkTest]
     public class ThreadProfileTests : IClassFixture<MvcWithCollectorFixture>
